services/controller/requirements.txt
websocket-client==1.8.0
rel==0.4.9.19
python-dotenv==1.0.1
paho-mqtt==2.1.0
